Moving index.js to /src
